requirements.txt
discord.py>=2.3.2
expr.py>=0.3.0